Add files via upload
This is an example of how you can easy create responsive layout to any of your website, using just grid method. 
Also, i made a overlay-grid, just to demonstrate how it's distributing underneath of your website, for visibility and better understanding of it.
You can toggle this overlay by chaking out a chackbox on the top of the page!